+++ src/core/utils/hooks/useTheme.ts
@@ -5,49 +5,65 @@
 import React from 'react';
 import '@itwin/itwinui-css/css/global.css';
 
+export type ThemeOptions = {
+  /**
+   * Document to which the theme will be applied.
+   * Can be specified to handle popup windows.
+   * @default document
+   */
+  ownerDocument?: Document;
+};
+
 export type ThemeType = 'light' | 'dark' | 'os';
 
 /**
  * Hook that applies styling and theme to all components.
  * Defaults to light theme if none provided or set elsewhere.
  * @param theme Light, dark, or based on OS setting.
+ * @param themeOptions Options that override default theming behavior.
  */
-export const useTheme = (theme?: ThemeType): void => {
+export const useTheme = (
+  theme?: ThemeType,
+  themeOptions?: ThemeOptions,
+): void => {
+  const ownerDocument = themeOptions?.ownerDocument ?? document;
   React.useLayoutEffect(() => {
-    if (!document.body.classList.contains('iui-body')) {
-      document.body.classList.add('iui-body');
+    if (!ownerDocument.body.classList.contains('iui-body')) {
+      ownerDocument.body.classList.add('iui-body');
     }
-  }, []);
+  }, [ownerDocument]);
 
   React.useLayoutEffect(() => {
     switch (theme) {
       case 'light':
-        addLightTheme();
+        addLightTheme(ownerDocument);
         break;
       case 'dark':
-        addDarkTheme();
+        addDarkTheme(ownerDocument);
         break;
       case 'os':
         if (window.matchMedia?.('(prefers-color-scheme: dark)').matches) {
-          addDarkTheme();
+          addDarkTheme(ownerDocument);
         } else {
-          addLightTheme();
+          addLightTheme(ownerDocument);
         }
         break;
       default:
-        if (!document.documentElement.classList.value.includes('iui-theme')) {
-          addLightTheme();
+        if (
+          !ownerDocument.documentElement.classList.value.includes('iui-theme')
+        ) {
+          addLightTheme(ownerDocument);
         }
     }
-  }, [theme]);
+  }, [ownerDocument, theme]);
 };
 
-const addLightTheme = () => {
-  document.documentElement.classList.add('iui-theme-light');
-  document.documentElement.classList.remove('iui-theme-dark');
+const addLightTheme = (ownerDocument: Document) => {
+  ownerDocument.documentElement.classList.add('iui-theme-light');
+  ownerDocument.documentElement.classList.remove('iui-theme-dark');
 };
 
-const addDarkTheme = () => {
-  document.documentElement.classList.add('iui-theme-dark');
-  document.documentElement.classList.remove('iui-theme-light');
+const addDarkTheme = (ownerDocument: Document) => {
+  ownerDocument.documentElement.classList.add('iui-theme-dark');
+  ownerDocument.documentElement.classList.remove('iui-theme-light');
 };